@@ -1,437 +1,437 @@
-export default [
+export const countries:Array<any> = [
 	{
-		label: "UK",
-		value: "GB"
+		label: 'UK',
+		value: 'GB'
 	},
 	{
-		label: "UK (N. IRELAND)",
-		value: "NB"
+		label: 'UK (N. IRELAND)',
+		value: 'NB'
 	},
 	{
-		label: "AUSTRIA",
-		value: "AT"
+		label: 'AUSTRIA',
+		value: 'AT'
 	},
 	{
-		label: "BELARUS",
-		value: "BY"
+		label: 'BELARUS',
+		value: 'BY'
 	},
 	{
-		label: "BELGIUM",
-		value: "BE"
+		label: 'BELGIUM',
+		value: 'BE'
 	},
 	{
-		label: "BULGARIA",
-		value: "BG"
+		label: 'BULGARIA',
+		value: 'BG'
 	},
 	{
-		label: "CANADA",
-		value: "CANADA",
+		label: 'CANADA',
+		value: 'CANADA',
 		states: [
 			{
-				"label": "AB",
-				"value": "AB"
+				'label': 'AB',
+				'value': 'AB'
 			},
 			{
-				"label": "BC",
-				"value": "BC"
+				'label': 'BC',
+				'value': 'BC'
 			},
 			{
-				"label": "MB",
-				"value": "MB"
+				'label': 'MB',
+				'value': 'MB'
 			},
 			{
-				"label": "NB",
-				"value": "NB"
+				'label': 'NB',
+				'value': 'NB'
 			},
 			{
-				"label": "NL",
-				"value": "NL"
+				'label': 'NL',
+				'value': 'NL'
 			},
 			{
-				"label": "NT",
-				"value": "NT"
+				'label': 'NT',
+				'value': 'NT'
 			},
 			{
-				"label": "NS",
-				"value": "NS"
+				'label': 'NS',
+				'value': 'NS'
 			},
 			{
-				"label": "NU",
-				"value": "NU"
+				'label': 'NU',
+				'value': 'NU'
 			},
 			{
-				"label": "ON",
-				"value": "ON"
+				'label': 'ON',
+				'value': 'ON'
 			},
 			{
-				"label": "PE",
-				"value": "PE"
+				'label': 'PE',
+				'value': 'PE'
 			},
 			{
-				"label": "QC",
-				"value": "QC"
+				'label': 'QC',
+				'value': 'QC'
 			},
 			{
-				"label": "SK",
-				"value": "SK"
+				'label': 'SK',
+				'value': 'SK'
 			},
 			{
-				"label": "YT",
-				"value": "YT"
+				'label': 'YT',
+				'value': 'YT'
 			}
 		]
 	},
 	{
-		label: "CROATIA",
-		value: "HR"
+		label: 'CROATIA',
+		value: 'HR'
 	},
 	{
-		label: "CZECH REPUBLIC",
-		value: "CZ"
+		label: 'CZECH REPUBLIC',
+		value: 'CZ'
 	},
 	{
-		label: "DENMARK",
-		value: "DK"
+		label: 'DENMARK',
+		value: 'DK'
 	},
 	{
-		label: "ESTONIA",
-		value: "EE"
+		label: 'ESTONIA',
+		value: 'EE'
 	},
 	{
-		label: "FINLAND",
-		value: "FI"
+		label: 'FINLAND',
+		value: 'FI'
 	},
 	{
-		label: "FRANCE",
-		value: "FR"
+		label: 'FRANCE',
+		value: 'FR'
 	},
 	{
-		label: "GERMANY",
-		value: "DE"
+		label: 'GERMANY',
+		value: 'DE'
 	},
 	{
-		label: "GREECE",
-		value: "GR"
+		label: 'GREECE',
+		value: 'GR'
 	},
 	{
-		label: "HUNGARY",
-		value: "HU"
+		label: 'HUNGARY',
+		value: 'HU'
 	},
 	{
-		label: "ICELAND",
-		value: "IS"
+		label: 'ICELAND',
+		value: 'IS'
 	},
 	{
-		label: "IRELAND",
-		value: "IE"
+		label: 'IRELAND',
+		value: 'IE'
 	},
 	{
-		label: "ITALY",
-		value: "IT"
+		label: 'ITALY',
+		value: 'IT'
 	},
 	{
-		label: "LATVIA",
-		value: "LV"
+		label: 'LATVIA',
+		value: 'LV'
 	},
 	{
-		label: "LITHUANIA",
-		value: "LT"
+		label: 'LITHUANIA',
+		value: 'LT'
 	},
 	{
-		label: "LUXEMBOURG",
-		value: "LU"
+		label: 'LUXEMBOURG',
+		value: 'LU'
 	},
 	{
-		label: "MONACO",
-		value: "MC"
+		label: 'MONACO',
+		value: 'MC'
 	},
 	{
-		label: "NETHERLANDS",
-		value: "NL"
+		label: 'NETHERLANDS',
+		value: 'NL'
 	},
 	{
-		label: "NORWAY",
-		value: "NO"
+		label: 'NORWAY',
+		value: 'NO'
 	},
 	{
-		label: "POLAND",
-		value: "PL"
+		label: 'POLAND',
+		value: 'PL'
 	},
 	{
-		label: "PORTUGAL",
-		value: "PT"
+		label: 'PORTUGAL',
+		value: 'PT'
 	},
 	{
-		label: "ROMANIA",
-		value: "RO"
+		label: 'ROMANIA',
+		value: 'RO'
 	},
 	{
-		label: "RUSSIA",
-		value: "RU"
+		label: 'RUSSIA',
+		value: 'RU'
 	},
 	{
-		label: "SLOVAKIA",
-		value: "SK"
+		label: 'SLOVAKIA',
+		value: 'SK'
 	},
 	{
-		label: "SLOVENIA",
-		value: "SI"
+		label: 'SLOVENIA',
+		value: 'SI'
 	},
 	{
-		label: "SPAIN",
-		value: "ES"
+		label: 'SPAIN',
+		value: 'ES'
 	},
 	{
-		label: "SWEDEN",
-		value: "SE"
+		label: 'SWEDEN',
+		value: 'SE'
 	},
 	{
-		label: "SWITZERLAND",
-		value: "CH"
+		label: 'SWITZERLAND',
+		value: 'CH'
 	},
 	{
-		label: "TURKEY",
-		value: "TR"
+		label: 'TURKEY',
+		value: 'TR'
 	}, 
 	{
-		label: "United States",
-		value: "USA",
+		label: 'United States',
+		value: 'USA',
 		states: [          
 			{
-				"label": "AL",
-				"value": "AL"
+				'label': 'AL',
+				'value': 'AL'
 			},
 			{
-				"label": "AK",
-				"value": "AK"
+				'label': 'AK',
+				'value': 'AK'
 			},
 			{
-				"label": "AS",
-				"value": "AS"
+				'label': 'AS',
+				'value': 'AS'
 			},
 			{
-				"label": "AZ",
-				"value": "AZ"
+				'label': 'AZ',
+				'value': 'AZ'
 			},
 			{
-				"label": "AR",
-				"value": "AR"
+				'label': 'AR',
+				'value': 'AR'
 			},
 			{
-				"label": "CA",
-				"value": "CA"
+				'label': 'CA',
+				'value': 'CA'
 			},
 			{
-				"label": "CO",
-				"value": "CO"
+				'label': 'CO',
+				'value': 'CO'
 			},
 			{
-				"label": "CT",
-				"value": "CT"
+				'label': 'CT',
+				'value': 'CT'
 			},
 			{
-				"label": "DE",
-				"value": "DE"
+				'label': 'DE',
+				'value': 'DE'
 			},
 			{
-				"label": "DC",
-				"value": "DC"
+				'label': 'DC',
+				'value': 'DC'
 			},
 			{
-				"label": "FM",
-				"value": "FM"
+				'label': 'FM',
+				'value': 'FM'
 			},
 			{
-				"label": "FL",
-				"value": "FL"
+				'label': 'FL',
+				'value': 'FL'
 			},
 			{
-				"label": "GA",
-				"value": "GA"
+				'label': 'GA',
+				'value': 'GA'
 			},
 			{
-				"label": "GU",
-				"value": "GU"
+				'label': 'GU',
+				'value': 'GU'
 			},
 			{
-				"label": "HI",
-				"value": "HI"
+				'label': 'HI',
+				'value': 'HI'
 			},
 			{
-				"label": "ID",
-				"value": "ID"
+				'label': 'ID',
+				'value': 'ID'
 			},
 			{
-				"label": "IL",
-				"value": "IL"
+				'label': 'IL',
+				'value': 'IL'
 			},
 			{
-				"label": "IN",
-				"value": "IN"
+				'label': 'IN',
+				'value': 'IN'
 			},
 			{
-				"label": "IA",
-				"value": "IA"
+				'label': 'IA',
+				'value': 'IA'
 			},
 			{
-				"label": "KS",
-				"value": "KS"
+				'label': 'KS',
+				'value': 'KS'
 			},
 			{
-				"label": "KY",
-				"value": "KY"
+				'label': 'KY',
+				'value': 'KY'
 			},
 			{
-				"label": "LA",
-				"value": "LA"
+				'label': 'LA',
+				'value': 'LA'
 			},
 			{
-				"label": "ME",
-				"value": "ME"
+				'label': 'ME',
+				'value': 'ME'
 			},
 			{
-				"label": "MH",
-				"value": "MH"
+				'label': 'MH',
+				'value': 'MH'
 			},
 			{
-				"label": "MD",
-				"value": "MD"
+				'label': 'MD',
+				'value': 'MD'
 			},
 			{
-				"label": "MA",
-				"value": "MA"
+				'label': 'MA',
+				'value': 'MA'
 			},
 			{
-				"label": "MI",
-				"value": "MI"
+				'label': 'MI',
+				'value': 'MI'
 			},
 			{
-				"label": "MN",
-				"value": "MN"
+				'label': 'MN',
+				'value': 'MN'
 			},
 			{
-				"label": "MS",
-				"value": "MS"
+				'label': 'MS',
+				'value': 'MS'
 			},
 			{
-				"label": "MO",
-				"value": "MO"
+				'label': 'MO',
+				'value': 'MO'
 			},
 			{
-				"label": "MT",
-				"value": "MT"
+				'label': 'MT',
+				'value': 'MT'
 			},
 			{
-				"label": "NE",
-				"value": "NE"
+				'label': 'NE',
+				'value': 'NE'
 			},
 			{
-				"label": "NV",
-				"value": "NV"
+				'label': 'NV',
+				'value': 'NV'
 			},
 			{
-				"label": "NH",
-				"value": "NH"
+				'label': 'NH',
+				'value': 'NH'
 			},
 			{
-				"label": "NJ",
-				"value": "NJ"
+				'label': 'NJ',
+				'value': 'NJ'
 			},
 			{
-				"label": "NM",
-				"value": "NM"
+				'label': 'NM',
+				'value': 'NM'
 			},
 			{
-				"label": "NY",
-				"value": "NY"
+				'label': 'NY',
+				'value': 'NY'
 			},
 			{
-				"label": "NC",
-				"value": "NC"
+				'label': 'NC',
+				'value': 'NC'
 			},
 			{
-				"label": "ND",
-				"value": "ND"
+				'label': 'ND',
+				'value': 'ND'
 			},
 			{
-				"label": "MP",
-				"value": "MP"
+				'label': 'MP',
+				'value': 'MP'
 			},
 			{
-				"label": "OH",
-				"value": "OH"
+				'label': 'OH',
+				'value': 'OH'
 			},
 			{
-				"label": "OK",
-				"value": "OK"
+				'label': 'OK',
+				'value': 'OK'
 			},
 			{
-				"label": "OR",
-				"value": "OR"
+				'label': 'OR',
+				'value': 'OR'
 			},
 			{
-				"label": "PW",
-				"value": "PW"
+				'label': 'PW',
+				'value': 'PW'
 			},
 			{
-				"label": "PA",
-				"value": "PA"
+				'label': 'PA',
+				'value': 'PA'
 			},
 			{
-				"label": "PR",
-				"value": "PR"
+				'label': 'PR',
+				'value': 'PR'
 			},
 			{
-				"label": "RI",
-				"value": "RI"
+				'label': 'RI',
+				'value': 'RI'
 			},
 			{
-				"label": "SC",
-				"value": "SC"
+				'label': 'SC',
+				'value': 'SC'
 			},
 			{
-				"label": "SD",
-				"value": "SD"
+				'label': 'SD',
+				'value': 'SD'
 			},
 			{
-				"label": "TN",
-				"value": "TN"
+				'label': 'TN',
+				'value': 'TN'
 			},
 			{
-				"label": "TX",
-				"value": "TX"
+				'label': 'TX',
+				'value': 'TX'
 			},
 			{
-				"label": "UT",
-				"value": "UT"
+				'label': 'UT',
+				'value': 'UT'
 			},
 			{
-				"label": "VT",
-				"value": "VT"
+				'label': 'VT',
+				'value': 'VT'
 			},
 			{
-				"label": "VI",
-				"value": "VI"
+				'label': 'VI',
+				'value': 'VI'
 			},
 			{
-				"label": "VA",
-				"value": "VA"
+				'label': 'VA',
+				'value': 'VA'
 			},
 			{
-				"label": "WA",
-				"value": "WA"
+				'label': 'WA',
+				'value': 'WA'
 			},
 			{
-				"label": "WV",
-				"value": "WV"
+				'label': 'WV',
+				'value': 'WV'
 			},
 			{
-				"label": "WI",
-				"value": "WI"
+				'label': 'WI',
+				'value': 'WI'
 			},
 			{
-				"label": "WY",
-				"value": "WY"
+				'label': 'WY',
+				'value': 'WY'
 			}
 		]
 	}
