@@ -1,10 +1,5 @@
-import {default as countries} from './countries';
+import { countries } from './countries';
 import * as discord from './discord';
 import * as sites from './sites';
 
-export {
-	countries,
-	discord,
-	//menuTemplate: require('./menu-template'),
-	sites
-};
+export default { countries, discord, sites };
