@@ -1,4 +1,4 @@
 import SupremeSafe from './SupremeSafe';
 import SupremeRequest from './SupremeRequest';
 
-export { SupremeSafe, SupremeRequest };
+export { SupremeRequest, SupremeSafe };
