@@ -1,6 +1,6 @@
 const electron = require('electron');
 const { BrowserWindow, session } = electron;
-const path = require('path')
+const path = require('path');
 
 function createWindow(accountName) {
 	const googleWindow = new BrowserWindow({
@@ -13,14 +13,15 @@ function createWindow(accountName) {
 			nodeIntegration: false,
 			preload: path.join(__dirname, 'preload.js')
 		}
-	})
+	});
 	module.exports.window = googleWindow;
 }
 
 async function loadWindow() {
 	const googleWindow = module.exports.window;
 	googleWindow.loadURL('https://accounts.google.com/ServiceLogin?service=youtube&uilel=3&passive=true&continue=https%3A%2F%2Fwww.youtube.com%2Fsignin%3Faction_handle_signin%3Dtrue%26app%3Dm%26hl%3Den-GB%26next%3Dhttps%253A%252F%252Fm.youtube.com%252F&hl=en-GB', {
-		userAgent: 'Mozilla/5.0 (X11; Ubuntu; Linux x86_64; rv:70.0) Gecko/20100101 Firefox/70.0'
+		// userAgent: 'Mozilla/5.0 (X11; Ubuntu; Linux x86_64; rv:70.0) Gecko/20100101 Firefox/70.0'
+		userAgent: 'Chrome'
 	});
 	googleWindow.show();
 }	
@@ -29,4 +30,4 @@ module.exports = {
 	create: createWindow,
 	load: loadWindow,
 	window: null
-}
+};
