@@ -1,9 +1,12 @@
 import './elements';
 import request from 'request-promise-native';
+import { RequestResponse } from 'request';
+
+
 
 export default function getProducts():Promise<any> {
 	return new Promise((resolve:Function):void => {
-		let productApiStatus:any = document.getElementById('productApiStatus');
+		let productApiStatus:HTMLInputElement = <HTMLInputElement>document.getElementById('productApiStatus');
 		productApiStatus.value = 'Fetching Products';
 		request({
 			url: 'http://prosper-products-eu.herokuapp.com/supreme/latest',
@@ -14,12 +17,12 @@ export default function getProducts():Promise<any> {
 				accept: 'application/json'
 			}
 		})
-			.then((response:any):void => {
+			.then((response:RequestResponse):void => {
 				let body:any = response.body;
 				productApiStatus.value = 'Loaded Products';
 				resolve(body);
 			})
-			.catch((error:any):void => {
+			.catch((error:Error):void => {
 				productApiStatus.value = 'Error Fetching Products';
 				console.log(error);
 				resolve({});
